+++ R/graticule.R
@@ -60,7 +60,7 @@
 #' }))
 #' plot(usa, graticule = st_crs(4326), axes = TRUE, lon = seq(-60,-130,by=-10))
 #' }
-st_graticule = function(x = c(-180,-90,180,90), crs = st_crs(x),
+st_graticule = function(x = c(-180, -90, 180, 90), crs = st_crs(x),
 	datum = st_crs(4326), ..., lon = NULL, lat = NULL, ndiscr = 100,
 	margin = 0.001)
 {