+++ tests/spatstat.R
@@ -50,7 +50,7 @@ as.ppp(sf) # warns
 w = st_as_sfc(st_bbox(st_sfc(p1, p2)))
 sf = st_sf(a = 1:3, geom = p)
 (p0 = rbind(st_sf(a = 0, geom = w), sf))
-try(as.ppp(p0)) # errors: one point outside window
+suppressWarnings(try(as.ppp(p0))) # errors: one point outside window
 
 w = st_as_sfc(st_bbox(p))
 sf = st_sf(a = 1:3, geom = p)