+++ tests/testthat/test-sfc.R
@@ -119,3 +119,11 @@ test_that("c.sfc n_empty returns sum of st_is_empty(sfg)", {
 test_that("st_is_longlat warns on invalid bounding box", {
 	expect_warning(st_is_longlat(st_sfc(st_point(c(0,-95)), crs = 4326)))
 })
+
+test_that("bounding box is flipped when geometry is flipped", {
+	foo <- st_bbox(c(xmin = 0, xmax = 100, ymin = 0, ymax = 200)) |> st_as_sfc()
+	bar <- foo * matrix(c(1,0,0,-1), nrow = 2)
+	expect_equal(st_bbox(bar), st_bbox(c(xmin=0, ymin=-200, xmax=100, ymax=0)))
+})
+
+