+++ NEWS.md
@@ -1,6 +1,22 @@
-# version 1.0-16
+# version 1.0-17
+
+* `st_sf()` no longer strips `tbl` or `tbl_df` class labels; #2378
+
+* `st_layers()` returns an object of class `c("sf_layers", "data.frame")`, with a dedicated `print` method.
+
+* when `dim` is not `XYZM`, `sf_as_sf.data.frame()` interprets a length 4 `coords` argument to specify the corners of a rectangular polygon; #2357
+
+* `st_interpolate_aw()` gains an `na.rm` argument, for removing features with `NA` attributes before interpolating; #830
 
 * `merge.sf()` no longer renames geometry column; #2334
+
+# version 1.0-16
+
+* `st_join()` no longer renames the geometry column; #2327
+
+* `st_sample()` works when unprojected polygon geometry crosses the antemeridian; #2331
+
+* clean up and modernization of S3 registration of methods and tests; #2285, #2288, #2316, #2341, #2342, by @olivroy
 
 * `[.sfc` works when setting argument `op`; #2320